requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
beautifulsoup4==4.12.3
django_channels==0.7.0
daphne==4.1.2
channels==4.1.0
channels_redis==4.2.0
python-dotenv==1.0.1